requirements.txt
Flask==3.0.0
Pillow==10.1.0
torch>=2.0.0
torchvision>=0.15.0
segment-anything
opencv-python>=4.8.0
numpy>=1.24.0

